requirements.txt
# OIE MVP - Requirements pentru Backtesting

# Core Backend
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data Fetching
aiohttp>=3.8.0

# Data Processing  
pandas>=2.0.0
numpy>=1.24.0

# Environment & Config
python-dotenv>=1.0.0

# Data Sources
yfinance>=0.2.0

# Opțional - pentru analiză avansată
# matplotlib>=3.7.0
# seaborn>=0.12.0
# ta>=0.10.0  # Technical Analysis library
